Moved metrics utils into a 'metrics' folder
