+++ .gitignore
@@ -1,3 +1,3 @@
-/appsody_reports
-/node_modules
+/metrics/appsody_reports
+/metrics/node_modules
 .DS_Store
