+++ metrics/data_pulling_scripts/appsody_stars_watchers_forks.js
@@ -1,5 +1,5 @@
 const { graphql } = require('@octokit/graphql')
-const tools = require('../data_pulling_utils/tools');
+const tools = require('../data_pulling_utils/data_pulling_tools');
 
 const args = process.argv.slice(2)
